+++ local_tests/testthat/test-getHUCName.R
@@ -80,9 +80,9 @@ testthat::test_that("allData returns are correct dimension and type", {
   spatialDataDir <- setup_HUCs()
 
   testthat::expect_s3_class(getHUCName(-114, 32, dataset = "WBDHU4_02", allData=TRUE), "data.frame")
-  testthat::expect_equal(dim(getHUCName(-100, 48, dataset = "WBDHU4_02", allData=TRUE)), c(1,ncol(WBDHU4_02)))
+  testthat::expect_equal(dim(getHUCName(-100, 48, dataset = "WBDHU4_02", allData=TRUE)), c(1,ncol(WBDHU4_02)-1))
   testthat::expect_s3_class(getHUCName(c(-80, -90), c(40, 41), dataset = "WBDHU4_02", allData=TRUE), "data.frame")
-  testthat::expect_equal(dim(getHUCName(c(-80, -90), c(40, 41), dataset = "WBDHU4_02", allData=TRUE)), c(2,ncol(WBDHU4_02)))
+  testthat::expect_equal(dim(getHUCName(c(-80, -90), c(40, 41), dataset = "WBDHU4_02", allData=TRUE)), c(2,ncol(WBDHU4_02)-1))
 
   # Teardown
   if (class(spatialDataDir) == "character") {