+++ NEWS.md
@@ -1,3 +1,19 @@
+# MazamaSpatialUtils 0.8.3
+
+* Added handling of "Sparse geometry binary predicate list of length 1" error in 
+`getSpatialData()` by simply choosing the first polygon returned.
+* Added `convertWBDHUC()` and pre-generated spatial data files: `WBDHU2`, 
+`WBDHU4`, `WBDHU6`.
+* Added `getHUC()` and `getHUCName()`.
+
+Added the following datasets to the archive at: 
+http://data.mazamascience.com/MazamaSpatialUtils/Spatial_0.8/
+* `WBDHU2` (and simplified versions)
+* `WBDHU4` (and simplified versions)
+* `WBDHU6` (and simplified versions)
+* `WBDHU8`
+* `WBDHU10`
+
 # MazamaSpatialUtils 0.8.2
 
 * Documentation fixes associated with change from **sp** to **sf**.